feat(compress): enable config the zstd compressor for encoding response
